+++ src/api/mod.rs
@@ -7,6 +7,7 @@ mod extensions;
 mod types;
 mod use_tiptap_editor;
 
+pub use editor::TiptapEditor;
 pub use editor::{TiptapEditorHandle, TiptapEditorInstance};
 pub use extensions::TiptapExtension;
 pub use types::{